+++ src/test/java/nrv/costs/dao/IncomeDaoTest.java
@@ -7,6 +7,8 @@ import org.junit.Assert;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
+import java.math.BigDecimal;
+
 
 public class IncomeDaoTest extends BaseTest {
 
@@ -16,7 +18,7 @@ public class IncomeDaoTest extends BaseTest {
 
     @Test
     public void testInsert() {
-        int id = dao.insert(new Income(83000.0, "Зарплата за июнь 2015"));
+        int id = dao.insert(new Income(BigDecimal.valueOf(83000.0), "Зарплата за июнь 2015"));
         Assert.assertNotNull("Just inserted id is null", id);
     }
 }
